pyproject.toml
[project]
name = "glupper"
version = "0.1.0"
description = "Trust-network backend for invite-based human verification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "asyncpg>=0.28.0",
    "pydantic[email]>=2.4.2",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "redis[hiredis]>=5.0.1",
    "mypy>=1.15.0",
    "ruff>=0.9.10",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "httpx>=0.24.1",
    "boto3>=1.37.11",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 160
target-version = "py312"
# Allow autofix for all enabled rules (when `--fix`) is provided.

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # not interested in docstring linting
    "D",
    "ERA001", # flag commented-out code
    "TD002", # missing issue link after TODO
    "TD003", # TODO should start with an uppercase word
    "T201", # flag print statements
    "PTH", # encourage use of Pathlib over os.path
    "UP007", # bad new union type annotation
    "PLR0913", # too many arguments to function def,
    "PD901", # dont call dfs df,
    "G004",
    "S311",
    "PLR2004", # magic value warnings,
    "FBT001", # allow bools in function params
    "S101", # asserts are okay
    "DTZ005", # absolutely do not want timezone aware timestamps
    "DTZ006", # same
    "FBT002",
    "FIX002", # flag TODO comments
]
fixable = ["A", "B", "C", "D", "I", "E", "F", "COM"]
unfixable = []

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
ignore_missing_imports = true
no_implicit_optional = true
no_implicit_reexport = true
show_error_codes = true
strict = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true

# set lint.mccabe for max-complexity?
